pyproject.toml
[tool.poetry]
name = "censys-cloud-connectors"
version = "3.1.0b1"
description = "The Censys Unified Cloud Connector is a standalone connector that gathers assets from various cloud providers and stores them in Censys ASM."
authors = ["Censys, Inc. <support@censys.io>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "censys/cloud_connectors", from = "src" }]
keywords = ["censys", "cloud", "connector", "attack surface management"]
classifiers = [
    "Typing :: Typed",
    "Topic :: Internet",
    "Topic :: Security",
    "Topic :: System :: Networking",
    "Topic :: System :: Networking :: Monitoring",
    "Topic :: Utilities",
    "Environment :: Console",
    "Natural Language :: English",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.urls]
"Censys Homepage" = "https://censys.io/"
"Discussions" = "https://github.com/censys/censys-cloud-connector/discussions"
"Changelog" = "https://github.com/censys/censys-cloud-connector/releases"
"Tracker" = "https://github.com/censys/censys-cloud-connector/issues"
"Source" = "https://github.com/censys/censys-cloud-connector"

[tool.poetry.scripts]
censys-cc = "censys.cloud_connectors.common.cli:main"

[tool.poetry.dependencies]
python = "^3.9"
backoff = "^2.1.2"
censys = "^2.1.8"
inquirerpy = "^0.3.3"
pydantic = {extras = ["dotenv", "email"], version = "^1.9.0"}
PyYAML = "^6.0"
requests = "^2.27.1"
rich = "^12.4.4"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
# Debug
ipython = "^8.1.0"
# Linting
black = "^22.1.0"
blacken-docs = "^1.12.1"
darglint = "^1.8.1"
flake8 = "^4.0.1"
flake8-black = "^0.3.2"
flake8-bugbear = "^22.3.23"
flake8-comprehensions = "^3.8.0"
flake8-docstrings = "^1.6.0"
flake8-isort = "^4.1.1"
flake8-pytest-style = "^1.6.0"
flake8-simplify = "^0.18.0"
isort = "^5.10.1"
pep8-naming = "^0.12.1"
pre-commit = "^2.17.0"
pyupgrade = "^2.31.1"
# Typing
mypy = "^0.942"
types-PyYAML = "^6.0.5"
types-requests = "^2.27.14"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
parameterized = "^0.8.1"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
pytest-datadir = "^1.3.1"
pytest-mock = "^3.7.0"
responses = "^0.21.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
furo = "^2022.6.21"
myst-parser = "^0.18.0"
Sphinx = "^5.1.1"
sphinx-copybutton = "^0.5.0"
sphinx-prompt = "^1.5.0"
sphinxcontrib-autoprogram = "^0.1.7"
sphinx-autobuild = "^2021.3.14"

[tool.poetry.group.aws.dependencies]
boto3 = "^1.24.63"
boto3-stubs = {extras = ["apigateway", "apigatewayv2", "ec2", "ecs", "elb", "elbv2", "rds", "route53", "route53domains", "s3", "sts"], version = "^1.24.63"}

[tool.poetry.group.azure.dependencies]
azure-cli = "^2.34.1"
azure-common = "^1.1.28"
azure-core = "^1.22.0"
azure-identity = "^1.7.1"
azure-mgmt-compute = ">=25.0.0,<25.1.0"
azure-mgmt-containerinstance = "^9.1.0"
azure-mgmt-core = "^1.3.0"
azure-mgmt-dns = "^8.0.0"
azure-mgmt-network = "^19.3.0"
azure-mgmt-resource = "20.0.0"
azure-mgmt-sql = "^3.0.1"
azure-mgmt-storage = "^19.1.0"
azure-storage-blob = "^12.9.0"
# Specified sys_platform to avoid issues with platform_system
pywin32 = {version = "^303", markers = "sys_platform == 'win32'"}

[tool.poetry.group.gcp.dependencies]
google-auth = "^2.6.0"
google-cloud-resource-manager = "^1.4.1"
google-cloud-securitycenter = "^1.9.0"

[tool.black]
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src", "tests"]
multi_line_output = 3
known_first_party = ["censys.cloud_connectors"]
known_censys = ["censys"]
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "CENSYS",
    "FIRSTPARTY",
    "LOCALFOLDER"
]

[tool.mypy]
exclude = [".venv", ".git"]
files = ["src"]
namespace_packages = true
plugins = ["pydantic.mypy"]
pretty = true
python_version = "3.9"
strict_optional = true
warn_redundant_casts = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    # Common
    "InquirerPy.*",
    "importlib_metadata",
    # Tests
    "parameterized",
    # GCP
    "google.oauth2",
    # Azure
    "azure.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "azure.*"
ignore_errors = true

[build-system]
requires = ["poetry>=1.2.0", "poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
